minor relocation
